Spring 6/Hibernate 6 migration: fix test failures to 94 remaining (from 1300+)
Key fixes in this batch:
- Revert GenericDaoHibernate.save() to simple persist/merge (no ID clearing)
- Fix Liquibase changeset HQL: acl → sharingACL.acl (embedded property path)
- Fix DocumentTagManagerTest: null fallback for AopProxyUtils proxy unwrapping
- Fix FileMetadataHibernate: sum() returns Number (not String) for String columns
- Fix DatabaseUsageByUserGroupByResult: usage field String→Number
- Fix DatabaseUsageByGroupGroupByResult: groupId BigInteger→Long
- Fix GroupDaoHibernateImpl: owner.username ordering via join
- Fix AbstractFormManagerImpl: explicit fieldFormDao.save() for immediate ID
- Fix AppConfigManagerImpl: explicit saveAppConfigElement() for immediate ID
- Fix SysAdminGroupsManagerImpl: updated for Long groupId

+++ src/main/java/com/researchspace/dao/DatabaseUsageByGroupGroupByResult.java
@@ -1,7 +1,6 @@
 package com.researchspace.dao;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.util.Comparator;
 
 /**
@@ -10,7 +9,7 @@ import java.util.Comparator;
  */
 public class DatabaseUsageByGroupGroupByResult {
 
-  public DatabaseUsageByGroupGroupByResult(BigInteger groupId, Double fileusage) {
+  public DatabaseUsageByGroupGroupByResult(Long groupId, Double fileusage) {
     super();
     this.groupId = groupId;
     this.fileusage = fileusage;
@@ -39,11 +38,11 @@ public class DatabaseUsageByGroupGroupByResult {
     super();
   }
 
-  public BigInteger getGroupId() {
+  public Long getGroupId() {
     return groupId;
   }
 
-  void setGroupId(BigInteger groupId) {
+  void setGroupId(Long groupId) {
     this.groupId = groupId;
   }
 
@@ -89,7 +88,7 @@ public class DatabaseUsageByGroupGroupByResult {
     return true;
   }
 
-  private BigInteger groupId;
+  private Long groupId;
 
   private Double fileusage = 0d;
